requirements.txt
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.14.0
pytube3==9.6.4
rsa==4.6
Telethon==1.16.4
typing-extensions==3.7.4.3
